feat add-book button, script, and CSS
Add CSS for the table and new-book feature.
Script adds form fields to new book, validates inputs, and adds it to
the list if valid.
Refactored table code to add a thead/tbody section for easier table
clearing for repopulation.
Renamed populateList() to updateTable().
Moved dummy books to myLibrary{} intialization.